接通产品证据评估、版本化研究报告与输入修正 (#99)
* docs: plan scope-bound step 7 evidence coordination

* docs: bind step 7 registry and control integration seams

* Expose bounded evidence graph queries through P0-09

* docs: record registered graph query verification

* docs: separate coordinator from scientific input assembly

* docs: include approved execution decision in closeout

* feat(web): add scope-bound assessment coordinator

* fix(web): reject ungrounded assessment completion claims

* docs: record reviewed assessment coordinator verification

* docs: pin scientific assembly roots and candidate resource scope

* docs: clarify byte-identical resource relocation exception

* docs: distinguish candidate data resource from registered schema

* docs: preserve local evidence bindings across portrait integration

* docs: separate scientific producer assembly from measured graph feedback

* docs: align relocated resource inventory and task handoffs

* docs: require current approval blockers and stale evidence display

* docs: retain committed-file policy gap as final integration gate

* docs: consolidate contributor guidance and end-to-end acceptance scope

* feat(web): assemble source-bound scientific assessment inputs

* fix(web): fully version producer-bound product cases

* docs: record producer completion and complete the scoped backend entry

* docs: retain exact assertion boundary for measured graph inputs

* docs: constrain measured domain routing to existing ownership

* docs: include assessment privacy contract in final alignment

* docs: preserve canonical prior profiles during graph append

* Clarify private assessment evidence aliases

* Bind measured assessment evidence to canonical graph feedback

* Record measured feedback evidence and freeze interface cleanup

* Connect scope-bound assessment portrait and private evidence projection

* Keep Step 7 plan focused on remaining acceptance

* Fix Step 7 acceptance evidence and lifecycle boundaries

* Preserve bounded and explicit assessment value projection

* Clarify executable assessment options at provider boundary

* test(web): preserve first stop event after input revision

* Record genuine Step 7 acceptance and remaining scientific gate

* Add version-bound private research report delivery

* Add confirmed conditional comparison and graft entry

* Connect native research evidence, versioned reports, and conditional inputs

* Bind private research report context to verified source measurements

* Complete shared research reports and dependency-scoped input revisions

* fix: preserve evidence continuity and readable research reports

* Preserve safe assessment rejection feedback on explicit resume

* Allow complete CI gates and reconcile bounded integration evidence

* Respect effective evidence revisions in research reports

* Align research report version dispatch and effective CSV evidence

* Reject retired evidence in legacy report verification

+++ src/bridge/tool_packages/p0_02_cell_state/method_adapter.py
@@ -507,7 +507,7 @@ def _read_matrix(path: Path) -> sparse.csr_matrix:
 
 
 def _log1p_cp10k(counts: sparse.csr_matrix) -> sparse.csr_matrix:
-    totals = np.asarray(counts.sum(axis=1)).ravel()
+    totals = np.asarray(counts.sum(axis=1), dtype=np.float64).ravel()
     scale = np.divide(10_000.0, totals, out=np.zeros_like(totals), where=totals > 0)
     matrix = (sparse.diags(scale) @ counts).astype(np.float32).tocsr()
     matrix.data = np.log1p(matrix.data)
